Fix Self destruct Issue in StakingNode.sol

+++ src/interfaces/IynETH.sol
@@ -5,7 +5,7 @@ import {IERC20} from "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 import {IERC20Permit} from "@openzeppelin/contracts/token/ERC20/extensions/IERC20Permit.sol";
 
 interface IynETH is IERC20 {
-    function withdrawETH(uint ethAmount) external;
+    function withdrawETH(uint256 ethAmount) external;
     function processWithdrawnETH() external payable;
     function receiveRewards() external payable;
     function updateDepositsPaused(bool paused) external;
